requirements.txt
# Web Framework
flask>=2.3.0
werkzeug>=2.3.0

# Scraping
selenium>=4.15.0
webdriver-manager>=4.0.0

# Telegram Bot
python-telegram-bot>=20.0

# Database (встроена в Python - sqlite3)
